IssueCommentReactionsClient

+++ Octokit/Clients/IIssueCommentReactionsClient.cs
@@ -19,7 +19,7 @@ namespace Octokit
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The comment id</param>
         /// <param name="reaction">The reaction to create</param>
-        Task<Reaction> Create(string owner, string name, int number, NewReaction reaction);
+        Task<Reaction> Create(string owner, string name, long number, NewReaction reaction);
 
         /// <summary>
         /// Creates a reaction for a specified Issue Comment
@@ -28,7 +28,7 @@ namespace Octokit
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The comment id</param>
         /// <param name="reaction">The reaction to create</param>
-        Task<Reaction> Create(long repositoryId, int number, NewReaction reaction);
+        Task<Reaction> Create(long repositoryId, long number, NewReaction reaction);
 
         /// <summary>
         /// Get all reactions for a specified Issue Comment
@@ -37,7 +37,7 @@ namespace Octokit
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The comment id</param>        
-        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, int number);
+        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, long number);
 
         /// <summary>
         /// Get all reactions for a specified Issue Comment
@@ -47,7 +47,7 @@ namespace Octokit
         /// <param name="name">The name of the repository</param>
         /// <param name="number">The comment id</param>
         /// <param name="options">Options for changing the API response</param>        
-        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, int number, ApiOptions options);
+        Task<IReadOnlyList<Reaction>> GetAll(string owner, string name, long number, ApiOptions options);
 
         /// <summary>
         /// Get all reactions for a specified Issue Comment
@@ -55,7 +55,7 @@ namespace Octokit
         /// <remarks>https://developer.github.com/v3/reactions/#list-reactions-for-an-issue-comment</remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The comment id</param>        
-        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, int number);
+        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, long number);
 
         /// <summary>
         /// Get all reactions for a specified Issue Comment
@@ -64,6 +64,6 @@ namespace Octokit
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="number">The comment id</param>
         /// <param name="options">Options for changing the API response</param>        
-        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, int number, ApiOptions options);
+        Task<IReadOnlyList<Reaction>> GetAll(long repositoryId, long number, ApiOptions options);
     }
 }
